Создал экземпляры компонентов в контроллере. Подключил контроллер в точку входа.
